+++ fiberplane-api-client/src/lib.rs
@@ -462,6 +462,23 @@ If you wish to delete a single key instead of the whole object, use the `patch` 
         Ok(response)
     }
 
+    #[doc = r#"Add a GitHub pull request to the notebooks front matter"#]
+    pub async fn integrations_github_app_pull_request_front_matter_add(
+        &self,
+        notebook_id: base64uuid::Base64Uuid,
+        payload: models::GitHubAppAddPullRequest,
+    ) -> Result<(), ApiClientError<models::GitHubAppAddPullRequestError>> {
+        let path = &format!(
+            "/api/notebooks/{notebookId}/integrations/github/pull_request",
+            notebookId = notebook_id,
+        );
+        let mut req = self.request(Method::POST, path)?;
+
+        let req = req.json(&payload);
+
+        self.do_req(req).await
+    }
+
     #[doc = r#"Convert the notebook cells to a snippet"#]
     pub async fn notebook_convert_to_snippet(
         &self,
@@ -570,7 +587,7 @@ For authenticating with the API see the Authentication section in the docs
         provider: &str,
         cli_redirect_port: Option<i32>,
         redirect: Option<&str>,
-    ) -> Result<models::OidLinkupLocation> {
+    ) -> Result<models::SoftRedirect> {
         let mut builder = self.request(
             Method::POST,
             &format!("/api/oidc/linkup/{provider}", provider = provider,),
@@ -1130,7 +1147,7 @@ For authenticating with the API see the Authentication section in the docs
     pub async fn integrations_github_app_install(
         &self,
         workspace_id: base64uuid::Base64Uuid,
-    ) -> Result<(), ApiClientError<models::GitHubAppInstallFlowError>> {
+    ) -> Result<models::SoftRedirect, ApiClientError<models::GitHubAppInstallFlowError>> {
         let path = &format!(
             "/api/workspaces/{workspaceId}/integrations/github/install",
             workspaceId = workspace_id,
@@ -1144,7 +1161,7 @@ For authenticating with the API see the Authentication section in the docs
     pub async fn integrations_github_app_uninstall(
         &self,
         workspace_id: base64uuid::Base64Uuid,
-    ) -> Result<(), ApiClientError<models::GitHubAppUninstallError>> {
+    ) -> Result<models::SoftRedirect, ApiClientError<models::GitHubAppUninstallError>> {
         let path = &format!(
             "/api/workspaces/{workspaceId}/integrations/github/uninstall",
             workspaceId = workspace_id,