+++ fiberplane-hooks/src/useKeyPressEvent.ts
@@ -1,4 +1,5 @@
-import { useEffect } from "react";
+import { useDocumentEventHandler } from "./useDocumentEventHandler";
+import { useHandler } from "./useHandler";
 
 /**
  * Hook adding a key press event listener to the document.
@@ -9,14 +10,11 @@ export function useKeyPressEvent(
   key: string,
   handler: (event: KeyboardEvent) => void,
 ) {
-  useEffect(() => {
-    const onKeyDown = (event: KeyboardEvent) => {
-      if (event.key === key) {
-        handler(event);
-      }
-    };
+  const onKeyDown = useHandler((event: KeyboardEvent) => {
+    if (event.key === key) {
+      handler(event);
+    }
+  });
 
-    document.addEventListener("keydown", onKeyDown);
-    return () => document.removeEventListener("keydown", onKeyDown);
-  }, [key, handler]);
+  useDocumentEventHandler("keydown", onKeyDown);
 }