+++ fiberplane-ui/src/components/index.ts
@@ -1,4 +1,7 @@
+export * from "./AttachedPopup";
 export * from "./Button";
 export * from "./Icon";
 export * from "./IconButton";
 export * from "./Input";
+export * from "./Select";
+export * from "./Tabs";