+++ fiberplane-ui/src/index.ts
@@ -1,1 +1,3 @@
 export * from "./components";
+
+export { cancelEvent, stopPropagation } from "./utils";